# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/barang": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Get list of barang",
                "description": "Display a listing of the resource.",
                "operationId": "28dbfd166522e0488753ad64bd0b55cd",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Barang"
                ],
                "summary": "Create a new barang",
                "description": "Store a newly created resource in storage.",
                "operationId": "2dc96475680f690140c0d9bb5a51aa5c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_kategori",
                                    "nama_barang",
                                    "berat",
                                    "dimensi"
                                ],
                                "properties": {
                                    "id_kategori": {
                                        "type": "integer"
                                    },
                                    "nama_barang": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "berat": {
                                        "type": "number"
                                    },
                                    "dimensi": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/barang/{id}": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Get a barang by ID",
                "description": "Display the specified resource.",
                "operationId": "643d8beb7bf164a62ffecdb20d06f35d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Barang"
                ],
                "summary": "Update a barang",
                "description": "Update the specified resource in storage.",
                "operationId": "e25f3a462cd8916bd3d911ef67d015c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_kategori": {
                                        "type": "integer"
                                    },
                                    "nama_barang": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "berat": {
                                        "type": "number"
                                    },
                                    "dimensi": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Barang"
                ],
                "summary": "Delete a barang",
                "description": "Remove the specified resource from storage.",
                "operationId": "e40430bacb2eee82718919d54df4a7f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Gunakan format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}